moving useRegistrars to api module
