+++ src/layout/RegistrarSelectionModal.tsx
@@ -2,7 +2,7 @@ import React, {useCallback, useState} from 'react';
 import BN from 'bn.js';
 import {StyleSheet, Dimensions, View} from 'react-native';
 import {Select, SelectItem, Modal, Card, Text, Button, IndexPath, Input, Icon, IconProps} from '@ui-kitten/components';
-import {useRegistrars} from 'src/hook/useRegistrars';
+import {useRegistrars} from 'src/api/hooks/useRegistrars';
 import {BN_ZERO} from '@polkadot/util';
 import {standardPadding, monofontFamily} from 'src/styles';
 import Identicon from '@polkadot/reactnative-identicon';
